3.1.4
- Improvement to multiple active timers: now stopping a timer will open
the last started timer in the plus menu and timer panel. Before, it was
opening one of the remaining active timers but it wasn't the last one
started. Now it has a consistent rule to which one to display next.

- performance improvements to coordinate timer panels with chrome
startup or trello page loading, and to coordinate opening multiple
active timers at once (when entering a board).

- temporarily stop doing chrome plus icon "rotation" animation until I
fix an issue that causes the icon to paint incorrectly or not update in
some cases.

- fix rare case when starting a timer could open two identical timer
panels.